debut mise en forme et debut ajout de la fonction testant plusieurs hyperparametres differents
